// src/libraries/Enum/url-enum.ts
export enum URL_ENUMS {
  BASE_URL = "http://api.openweathermap.org/",
  API_KEY = "1c5da32bd6a0d1c4c017b21b49833c7f",
  ICON_URL = " https://openweathermap.org/img/wn/",
  //geocoding
  GEOCODING_URL = `geo/1.0/direct`,
  WEATHER_FORECAST = "data/2.5/onecall",
  AIR_QUALITY = "data/2.5/air_pollution",
}